+++ packages/electron-adapter/src/Logger.ts
